+++ addon/components/item-picker/item-preview/component.js
@@ -1,9 +1,12 @@
-import Ember from 'ember';
+import { computed } from '@ember/object';
+import { notEmpty, reads } from '@ember/object/computed';
+import { inject as service } from '@ember/service';
+import Component from '@ember/component';
 import layout from './template';
 
-export default Ember.Component.extend({
+export default Component.extend({
   layout,
-  intl: Ember.inject.service(),
+  intl: service(),
 
   classNames: [ 'item-picker-current-item-preview' ],
 
@@ -17,13 +20,13 @@ export default Ember.Component.extend({
   isValidating: false,
   selectAnyway: false,
   shouldValidate: false,
-  showError: Ember.computed.notEmpty('errorMessage'),
-  description: Ember.computed.reads('model.description'),
+  showError: notEmpty('errorMessage'),
+  description: reads('model.description'),
 
   /**
    * What should the select button text be? we have variations depending on status
    */
-  selectButtonText: Ember.computed('isValidating', 'selectAnyway', function () {
+  selectButtonText: computed('isValidating', 'selectAnyway', function () {
     const intl = this.get('intl');
     let key = 'buttons.select';
     if (this.get('isValidating')) {
@@ -37,7 +40,7 @@ export default Ember.Component.extend({
   /**
    * Get the translated form of the Item Type
    */
-  itemType: Ember.computed('_i18nScope', 'model.type', function () {
+  itemType: computed('_i18nScope', 'model.type', function () {
     const itemType = this.get('model.type');
     let result = itemType;
     const key = `${this.get('_i18nScope')}shared.itemType.${itemType.camelize()}`;
@@ -52,7 +55,7 @@ export default Ember.Component.extend({
   /**
    * Construct the preview url
    */
-  previewUrl: Ember.computed('model', function () {
+  previewUrl: computed('model', function () {
     const item = this.get('model');
     let previewURL;
     // if the item has a url property, use that...
@@ -76,7 +79,7 @@ export default Ember.Component.extend({
   /**
    * What class should be used for any messages
    */
-  messageClass: Ember.computed('errorMessage', function () {
+  messageClass: computed('errorMessage', function () {
     if (this.get('errorMessage.status') === 'warning') {
       return 'alert-warning';
     } else if (this.get('errorMessage.status') === 'error') {